pyproject.toml
[project]
name = "voicebot"
version = "0.1.0"
description = "Discord bot that sends notifications whenever a user enters a voice channel"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py[voice]>=2.5.2",
    "dotenv>=0.9.9",
]
